fix(data-dashboard): handle no profile accounts instead of inifinte loading
